refactor: ActiveIndicatorContainer を独立させた
